+++ src/subsearch/gui/assets/buttons/exit.png
