+++ src/subsearch/gui/assets/buttons/exit_grey.png
