+++ src/subsearch/gui/assets/buttons/maximize.png
