+++ src/subsearch/gui/assets/buttons/maximize_grey.png
