+++ src/subsearch/gui/assets/buttons/maximize_inactive.png
