+++ src/subsearch/gui/assets/icons/128.ico
