+++ src/subsearch/gui/assets/icons/16.ico
