+++ src/subsearch/gui/assets/icons/256.ico
