+++ src/subsearch/gui/assets/icons/32.ico
