+++ src/subsearch/gui/assets/icons/64.ico
