+++ src/subsearch/gui/tkinter_data.py
