+++ src/subsearch/gui/widget_download.py
@@ -2,10 +2,10 @@ import tkinter as tk
 from tkinter import ttk
 
 import sv_ttk
-from src.gui import tkinter_data as tkd
-from src.gui import widget_root
-from src.scraper import subscene_soup
-from src.utilities import file_manager, local_paths
+from gui import tkinter_data as tkd
+from gui import widget_root
+from scraper import subscene_soup
+from util import file_manager, local_paths
 
 
 # file with subtitles and corresponding dl links
