+++ src/subsearch/gui/widget_root.py
@@ -1,7 +1,7 @@
 import tkinter as tk
 
-from src.gui import tools, tkinter_data as tkd
-from src.utilities import local_paths
+from gui import tools, tkinter_data as tkd
+from util import local_paths
 
 
 def main():
